# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  model.Age_Category:
    properties:
      icon_url:
        example: https://example.com/age-icon.png
        type: string
      id:
        example: 1
        type: integer
      range:
        example: 10-12
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Server for Ozinse Video Platform Backend
  title: Ozinse API
  version: "1.0"
paths:
  /api/age-categories:
    get:
      consumes:
      - application/json
      description: Returns all age_categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Age_Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List age_categories
      tags:
      - age_categories
    post:
      consumes:
      - application/json
      description: Creates a new age_category. Requires administrator role.
      parameters:
      - description: Age_Category payload
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.Age_Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Age_Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create age_category (Admin only)
      tags:
      - age_categories
  /api/age-categories/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an age_category by ID. Requires administrator role.
      parameters:
      - description: Age_Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete age_category (Admin only)
      tags:
      - age_categories
    get:
      consumes:
      - application/json
      description: Returns an age_category by its ID
      parameters:
      - description: Age_Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Age_Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get age_category by ID
      tags:
      - age_categories
    put:
      consumes:
      - application/json
      description: Updates an existing age_category by ID. Requires administrator
        role.
      parameters:
      - description: Age_Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Age_Category payload
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.Age_Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Age_Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update age_category (Admin only)
      tags:
      - age_categories
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"