# internal/handler/age_category_handler.go
package handler

import (
	"net/http"
	"ozinse-backend/internal/logger"
	"ozinse-backend/internal/model"
	"ozinse-backend/internal/service"
	"strconv"

	"github.com/gin-gonic/gin"
)

type AgeCategoryHandler struct {
	service *service.AgeCategoryService
}

func NewAgeCategoryHandler(service *service.AgeCategoryService) *AgeCategoryHandler {
	return &AgeCategoryHandler{service: service}
}

// GetAll age_categories
// @Summary List age_categories
// @Description Returns all age_categories
// @Tags age_categories
// @Accept json
// @Produce json
// @Success 200 {array} model.Age_Category
// @Failure 500 {object} map[string]string
// @Router /api/age-categories [get]
func (h *AgeCategoryHandler) GetAll(c *gin.Context) {
	// Define structured context fields for tracking the request
	log := logger.Log.With(
		"requestType", "GET",
		"endpoint", "/api/age-categories",
	)

	age_categories, err := h.service.GetAll()
	if err != nil {
		// Log persistent internal server issues
		log.Error("failed to retrieve all age categories from database", "error", err.Error())
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	// Log a successful bulk read operation
	log.Info("successfully fetched all age categories", "count", len(age_categories))
	c.JSON(http.StatusOK, age_categories)
}

// GetByID age_category by ID
// @Summary Get age_category by ID
// @Description Returns an age_category by its ID
// @Tags age_categories
// @Accept json
// @Produce json
// @Param id path int true "Age_Category ID"
// @Success 200 {object} model.Age_Category
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Router /api/age-categories/{id} [get]  
func (h *AgeCategoryHandler) GetByID(c *gin.Context) {
	idParam := c.Param("id")
	log := logger.Log.With(
		"requestType", "GET",
		"endpoint", "/api/age-categories/"+idParam,
	)

	id, err := strconv.Atoi(idParam)
	if err != nil {
		// Log syntax or mapping violations in request URI path parameters
		log.Warn("invalid age category ID path parameter structure", "raw_id", idParam)
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid id"})
		return
	}

	age_category, err := h.service.GetByID(id)
	if err != nil {
		// Log dynamic search failures cleanly as a system warning
		log.Warn("targeted age category record missing from storage layers", "age_category_id", id)
		c.JSON(http.StatusNotFound, gin.H{"error": "age_category with id " + idParam + " not found"})
		return
	}

	// Log successful identification showing relevant domain details
	log.Info("successfully found single age category record", "age_category_id", id, "range", age_category.Range)
	c.JSON(http.StatusOK, age_category)
}

// Create age_category
// @Summary Create age_category (Admin only)
// @Description Creates a new age_category. Requires administrator role.
// @Tags age_categories
// @Security BearerAuth  
// @Accept json
// @Produce json
// @Param category body model.Age_Category true "Age_Category payload"
// @Success 201 {object} model.Age_Category
// @Failure 401 {object} map[string]string "Unauthorized access"
// @Failure 403 {object} map[string]string "Forbidden - Admin only"
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/age-categories [post] 
func (h *AgeCategoryHandler) Create(c *gin.Context) {
	log := logger.Log.With(
		"requestType", "POST",
		"endpoint", "/api/age-categories",
	)

	var age_category model.Age_Category
	if err := c.ShouldBindJSON(&age_category); err != nil {
		// Log tracking data schema issues or blank body properties
		log.Error("failed to unpack structural validation formats for incoming payload", "error", err.Error())
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	if err := h.service.Create(&age_category); err != nil {
		// Log logical backend persistence layer rejections
		log.Error("failed to commit fresh age category creation transactions", "range", age_category.Range, "error", err.Error())
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	// Log execution success indicating tracking indexes generated by SQL
	log.Info("new age category committed and generated into runtime system", "generated_id", age_category.ID, "range", age_category.Range)
	c.JSON(http.StatusCreated, age_category)
}

// Update age_category
// @Summary Update age_category (Admin only)
// @Description Updates an existing age_category by ID. Requires administrator role.
// @Tags age_categories
// @Security BearerAuth  
// @Accept json
// @Produce json
// @Param id path int true "Age_Category ID"
// @Param category body model.Age_Category true "Age_Category payload"
// @Success 200 {object} model.Age_Category
// @Failure 401 {object} map[string]string "Unauthorized access"
// @Failure 403 {object} map[string]string "Forbidden - Admin only"
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/age-categories/{id} [put] 
func (h *AgeCategoryHandler) Update(c *gin.Context) {
	idParam := c.Param("id")
	log := logger.Log.With(
		"requestType", "PUT",
		"endpoint", "/api/age-categories/"+idParam,
	)

	id, err := strconv.Atoi(idParam)
	if err != nil {
		log.Warn("invalid age category ID submitted for transformation", "raw_id", idParam)
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid id"})
		return
	}

	var age_category model.Age_Category
	if err := c.ShouldBindJSON(&age_category); err != nil {
		log.Error("failed to translate structured update payload fields into entity definitions", "age_category_id", id, "error", err.Error())
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	age_category.ID = id
	if err := h.service.Update(&age_category); err != nil {
		log.Error("failed to sync requested resource property alterations to storage systems", "age_category_id", id, "error", err.Error())
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	log.Info("age category parameters successfully altered and updated", "age_category_id", id, "updated_range", age_category.Range)
	c.JSON(http.StatusOK, age_category)
}

// Delete age_category
// @Summary Delete age_category (Admin only)
// @Description Deletes an age_category by ID. Requires administrator role.
// @Tags age_categories
// @Security BearerAuth  
// @Accept json
// @Produce json
// @Param id path int true "Age_Category ID"
// @Success 200 {object} map[string]string "successfully deleted"
// @Failure 401 {object} map[string]string "Unauthorized access"
// @Failure 403 {object} map[string]string "Forbidden - Admin only"
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Router /api/age-categories/{id} [delete]
func (h *AgeCategoryHandler) Delete(c *gin.Context) {
	idParam := c.Param("id")
	log := logger.Log.With(
		"requestType", "DELETE",
		"endpoint", "/api/age-categories/"+idParam,
	)

	id, err := strconv.Atoi(idParam)
	if err != nil {
		log.Warn("invalid conversion structure for destruction target index parameters", "raw_id", idParam)
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid id"})
		return
	}

	if err := h.service.Delete(id); err != nil {
		log.Error("failed to remove existing target relational entity indexes", "age_category_id", id, "error", err.Error())
		c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
		return
	}

	log.Info("age category dropped and terminated from records systems completely", "deleted_id", id)
	c.JSON(http.StatusOK, gin.H{"message": "successfully deleted"})
}